feat: Add range proof splitting to handle large block intervals
- Add RANGE_PROOF_INTERVAL config (default 512 blocks) to split large proposals
- Implement sequential range proof generation with aggregation
- Prevents SP1 16 MiB witness limit errors for large block ranges
- Uses existing block range splitting utilities from cost estimator

Additional improvements:
- Optimize Rollup.sol struct packing (uint128->uint32 for block numbers)
- Change L1 head reference from block.number-5 to block.number-1
- Improve error handling in cost estimator with panic recovery
- Add detailed logging for RPC and rollup configuration
- Fix type conversions for new uint32 fields

+++ fault-proof/src/lib.rs
@@ -304,7 +304,7 @@ where
                 .header
                 .timestamp;
             
-            if proposal.deadline < current_timestamp {
+            if u64::from(proposal.deadline) < current_timestamp {
                 tracing::debug!(
                     "Proposal {} deadline {} has passed, cannot challenge/defend",
                     proposal_id,
